apps/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
databases[postgresql]>=0.6.0
asyncpg>=0.27.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
stripe>=5.0.0
rq>=1.10.0
sentry-sdk>=1.18.0
prometheus-fastapi-instrumentator>=6.0.0
python-dotenv>=1.0.0
pytest-asyncio>=0.21.0
factory-boy>=3.2.0
alembic
psycopg2-binary>=2.9.0
PyJWT
python-multipart